requirements.txt
Django==4.2.7
graphene-django==3.1.5
django-graphql-jwt==0.4.0
django-cors-headers==4.3.0
PyJWT==2.8.0
